+++ crates/node/src/commands/run.rs
@@ -1,13 +1,52 @@
 use std::path::PathBuf;
 
+use clap::ValueHint::FilePath;
 use madara_runtime::SealingMode;
 use mc_data_availability::DaLayer;
+use mc_l1_messages::config::{L1MessagesWorkerConfig, L1MessagesWorkerConfigError};
+use mc_settlement::SettlementLayer;
 use sc_cli::{Result, RpcMethods, RunCmd, SubstrateCli};
 use sc_service::BasePath;
 use serde::{Deserialize, Serialize};
 
 use crate::cli::Cli;
 use crate::service;
+
+#[derive(Debug, Clone, clap::Args)]
+#[group(multiple = true)]
+pub struct L1MessagesParams {
+    /// Ethereum Provider (Node) Url
+    #[clap(
+        long,
+        value_hint=clap::ValueHint::Url,
+        conflicts_with="l1_messages_config",
+        requires="l1_contract_address",
+    )]
+    pub provider_url: Option<String>,
+
+    /// L1 Contract Address
+    #[clap(
+        long,
+        value_hint=clap::ValueHint::Other,
+        conflicts_with="l1_messages_config",
+        requires="provider_url",
+    )]
+    pub l1_contract_address: Option<String>,
+}
+
+#[derive(Debug, Clone, clap::Args)]
+pub struct L1Messages {
+    /// Path to configuration file for Ethereum Core Contract Events Listener
+    #[clap(
+        long,
+        conflicts_with_all=["provider_url", "l1_contract_address"],
+        value_hint=clap::ValueHint::FilePath,
+    )]
+    pub l1_messages_config: Option<PathBuf>,
+
+    #[clap(flatten)]
+    pub config_params: L1MessagesParams,
+}
 
 /// Available Sealing methods.
 #[derive(Debug, Copy, Clone, clap::ValueEnum, Default, Serialize, Deserialize)]
@@ -42,11 +81,20 @@ pub struct ExtendedRunCmd {
     pub sealing: Option<Sealing>,
 
     /// Choose a supported DA Layer
-    #[clap(long, requires = "da_conf")]
+    #[clap(long, ignore_case = true, requires = "da_conf")]
     pub da_layer: Option<DaLayer>,
 
-    #[clap(long, requires = "da_layer")]
-    pub da_conf: Option<String>,
+    /// Path to a file containing the DA configuration
+    #[clap(long, value_hint = FilePath, requires = "da_layer")]
+    pub da_conf: Option<PathBuf>,
+
+    /// Choose a supported settlement layer
+    #[clap(long, ignore_case = true, requires = "settlement_conf")]
+    pub settlement: Option<SettlementLayer>,
+
+    /// Path to a file containing the settlement configuration
+    #[clap(long, value_hint = FilePath, requires = "settlement")]
+    pub settlement_conf: Option<PathBuf>,
 
     /// When enabled, more information about the blocks and their transaction is cached and stored
     /// in the database.
@@ -55,6 +103,10 @@ pub struct ExtendedRunCmd {
     /// increases the memory footprint of the node.
     #[clap(long)]
     pub cache: bool,
+
+    /// Configuration for L1 Messages (Syncing) Worker
+    #[clap(flatten)]
+    pub l1_messages_worker: L1Messages,
 }
 
 impl ExtendedRunCmd {
@@ -75,7 +127,7 @@ pub fn run_node(mut cli: Cli) -> Result<()> {
 
     let da_config: Option<(DaLayer, PathBuf)> = match cli.run.da_layer {
         Some(da_layer) => {
-            let da_conf = PathBuf::from(cli.run.da_conf.expect("clap requires da_conf when da_layer is present"));
+            let da_conf = cli.run.da_conf.expect("clap requires da_conf when da_layer is present");
             if !da_conf.exists() {
                 log::info!("{} does not contain DA config", da_conf.display());
                 return Err("DA config not available".into());
@@ -88,11 +140,49 @@ pub fn run_node(mut cli: Cli) -> Result<()> {
             None
         }
     };
+
+    let l1_messages_worker_config = extract_l1_messages_worker_config(&cli.run.l1_messages_worker)
+        .map_err(|e| sc_cli::Error::Input(e.to_string()))?;
+
+    let settlement_config: Option<(SettlementLayer, PathBuf)> = match cli.run.settlement {
+        Some(SettlementLayer::Ethereum) => {
+            let settlement_conf = cli.run.settlement_conf.expect("clap requires da_conf when settlement is present");
+            if !settlement_conf.exists() {
+                return Err(sc_cli::Error::Input(format!(
+                    "Ethereum config not found at {}",
+                    settlement_conf.display()
+                )));
+            }
+            Some((SettlementLayer::Ethereum, settlement_conf))
+        }
+        None => {
+            log::info!("Madara initialized w/o settlement layer");
+            None
+        }
+    };
+
     runner.run_node_until_exit(|config| async move {
         let sealing = cli.run.sealing.map(Into::into).unwrap_or_default();
         let cache = cli.run.cache;
-        service::new_full(config, sealing, da_config, cache).map_err(sc_cli::Error::Service)
+        service::new_full(config, sealing, da_config, cache, l1_messages_worker_config, settlement_config)
+            .map_err(sc_cli::Error::Service)
     })
+}
+
+fn extract_l1_messages_worker_config(
+    run_cmd: &L1Messages,
+) -> std::result::Result<Option<L1MessagesWorkerConfig>, L1MessagesWorkerConfigError> {
+    if let Some(ref config_path) = run_cmd.l1_messages_config {
+        let config = L1MessagesWorkerConfig::new_from_file(config_path)?;
+        return Ok(Some(config));
+    }
+
+    if let L1MessagesParams { provider_url: Some(url), l1_contract_address: Some(address) } = &run_cmd.config_params {
+        let config = L1MessagesWorkerConfig::new_from_params(url, address)?;
+        return Ok(Some(config));
+    }
+
+    Ok(None)
 }
 
 fn override_dev_environment(cmd: &mut ExtendedRunCmd) {