+++ crates/pallets/starknet/src/tests/declare_tx.rs
@@ -312,34 +312,6 @@ fn test_verify_tx_longevity() {
 }
 
 #[test]
-fn test_verify_no_require_tag() {
-    new_test_ext::<MockRuntime>().execute_with(|| {
-        basic_test_setup(2);
-
-        let chain_id = Starknet::chain_id();
-        let transaction =
-            get_declare_dummy(chain_id, Felt252Wrapper::ZERO, AccountType::V0(AccountTypeV0Inner::NoValidate));
-        let erc20_class = get_contract_class("ERC20.json", 0);
-
-        let validate_result = Starknet::validate_unsigned(
-            TransactionSource::InBlock,
-            &crate::Call::declare { transaction: transaction.clone(), contract_class: erc20_class },
-        )
-        .unwrap();
-
-        let valid_transaction_expected = ValidTransaction::with_tag_prefix("starknet")
-            .priority(u64::MAX - (TryInto::<u64>::try_into(*transaction.nonce())).unwrap())
-            .and_provides((*transaction.sender_address(), *transaction.nonce()))
-            .longevity(TransactionLongevity::get())
-            .propagate(true)
-            .build()
-            .unwrap();
-
-        assert_eq!(validate_result, valid_transaction_expected)
-    });
-}
-
-#[test]
 fn test_verify_require_tag() {
     new_test_ext::<MockRuntime>().execute_with(|| {
         basic_test_setup(2);
@@ -356,7 +328,7 @@ fn test_verify_require_tag() {
         .unwrap();
 
         let valid_transaction_expected = ValidTransaction::with_tag_prefix("starknet")
-            .priority(u64::MAX - (TryInto::<u64>::try_into(*transaction.nonce())).unwrap())
+            .priority(u64::MAX)
             .and_provides((*transaction.sender_address(), *transaction.nonce()))
             .longevity(TransactionLongevity::get())
             .propagate(true)