+++ madara-test-runner/src/client.rs
@@ -1,0 +1,2 @@
+// TODO: refactor starknet-rpc-test
+pub use starknet_rpc_test::MadaraClient;